+++ Alignment/OfflineValidation/python/TkAlAllInOneTool/configTemplates.py
@@ -17,6 +17,16 @@ from TkAlExceptions import AllInOneError
 
 ######################################################################
 ######################################################################
+loadGlobalTagTemplate="""
+#Global tag
+process.load("Configuration.StandardSequences.FrontierConditions_GlobalTag_cff")
+from Configuration.AlCa.GlobalTag import GlobalTag
+process.GlobalTag = GlobalTag(process.GlobalTag,".oO[GlobalTag]Oo.")
+"""
+
+
+######################################################################
+######################################################################
 conditionsTemplate="""
 process.conditionsIn.oO[rcdName]Oo. = CalibTracker.Configuration.Common.PoolDBESSource_cfi.poolDBESSource.clone(
      connect = cms.string('.oO[connectString]Oo.'),
@@ -200,38 +210,6 @@ else
     echo ".oO[validationId]Oo._result.root is up-to-date, no need to compare again."
     xrdcp -f root://eoscms//eos/cms/store/caf/user/$USER/.oO[eosdir]Oo./.oO[validationId]Oo._result.root .
 fi
-"""
-
-
-######################################################################
-######################################################################
-extendedValidationExecution="""
-#run extended offline validation scripts
-echo -e "\n\nRunning extended offline validation"
-
-rfcp .oO[extendedValScriptPath]Oo. .
-root -x -b -q -l TkAlExtendedOfflineValidation.C
-
-"""
-
-
-######################################################################
-######################################################################
-extendedValidationTemplate="""
-#include ".oO[CMSSW_BASE]Oo./src/Alignment/OfflineValidation/macros/PlotAlignmentValidation.C"
-void TkAlExtendedOfflineValidation()
-{
-  // load framework lite just to find the CMSSW libs...
-  gSystem->Load("libFWCoreFWLite");
-  FWLiteEnabler::enable();
-
-  .oO[extendedInstantiation]Oo.
-  p.setOutputDir(".oO[datadir]Oo./ExtendedOfflineValidation_Images");
-  p.setTreeBaseDir(".oO[OfflineTreeBaseDir]Oo.");
-  p.plotDMR(".oO[DMRMethod]Oo.",.oO[DMRMinimum]Oo.,".oO[DMROptions]Oo.");
-  p.plotSurfaceShapes(".oO[SurfaceShapes]Oo.");
-  p.plotChi2("root://eoscms//eos/cms/store/caf/user/$USER/.oO[eosdir]Oo./.oO[resultPlotFile]Oo._result.root");
-}
 """
