+++ Alignment/OfflineValidation/python/TkAlAllInOneTool/geometryComparisonTemplates.py
@@ -5,9 +5,7 @@ import FWCore.ParameterSet.Config as cms
 
 process = cms.Process("ValidationIntoNTuples")
 
-# global tag
-process.load("Configuration.StandardSequences.FrontierConditions_GlobalTag_cff")
-process.GlobalTag.globaltag = ".oO[GlobalTag]Oo." 
+.oO[LoadGlobalTagTemplate]Oo.
 
 process.load("Configuration.Geometry.GeometryRecoDB_cff")
 
@@ -43,9 +41,7 @@ import FWCore.ParameterSet.Config as cms
 
 process = cms.Process("validation")
 
-# global tag
-process.load("Configuration.StandardSequences.FrontierConditions_GlobalTag_cff")
-process.GlobalTag.globaltag = ".oO[GlobalTag]Oo." 
+.oO[LoadGlobalTagTemplate]Oo.
 
 process.load("Configuration.Geometry.GeometryRecoDB_cff")
